+++ CHANGELOG.md
@@ -4,8 +4,8 @@
 - Task Execution: Cancelled samples are now logged in the same fashion as samples with errors.
 - Sandboxes: Don't require Docker compatible sandboxes to implement `config_deserialize()`.
 - Limits: New `cost_limit()` context manager for scoped application of cost limits (can also be applied directly samples or agents).
-- Inspect View: Make samples in task detail sortable, inline epoch filter.
-- Inspect View: Show samples status.
+- Inspect View: Make samples in task detail sortable, inline epoch filter, show sample status.
+- Sandboxes: New `exec_remote()` method on `SandboxEnvironment` for asynchronous execution of long-running commands with streaming output.
 
 ## 0.3.179 (12 February 2026)